# docs/openapi.yaml
openapi: 3.1.0
info:
  title: outplan API
  version: "1.0.0"
  description: |
    Hand-written and kept in step with the code by hand. A change to an endpoint,
    a field, a status code or an error is not finished until this file says so.

    Everything under /api/v1 needs an Authorization header. Everything except
    POST /sessions also needs x-session-id, unless the credential is an agent key,
    which is itself the session.
servers:
  - url: https://api.outplan.org
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: A Firebase ID token, or an agent key beginning `oak_`.
    session:
      type: apiKey
      in: header
      name: x-session-id
      description: The device session. Omitted when the bearer is an agent key.
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, enum: [false] }
        error:
          type: object
          properties:
            code: { type: string, example: BAD_REQUEST }
            message: { type: string, description: Written to be shown to a person. }
            detail:
              type: object
              description: On validation failures, `fields` says which field and why.
    ReplyParts:
      type: object
      properties:
        messages:
          type: array
          maxItems: 3
          items: { type: string }
          description: One to three bubbles, the way someone types on a phone.
        question:
          type: object
          description: Dropped entirely if it has fewer than two options; one option is not a choice.
          properties:
            text: { type: string }
            options: { type: array, minItems: 2, maxItems: 4, items: { type: string } }
    Exchange:
      type: object
      properties:
        id: { type: string, format: uuid }
        said: { type: string }
        replied: { type: string, nullable: true }
        replyParts: { $ref: '#/components/schemas/ReplyParts' }
        coveredById:
          type: string
          format: uuid
          nullable: true
          description: >
            Set when this message was answered as part of a later reply, because
            several arrived before any answer. Not unanswered; answered together.
        saidAt: { type: string, format: date-time }
        repliedAt: { type: string, format: date-time, nullable: true }
        localDay: { type: string, example: "2026-09-18" }
        corrections: { type: array, items: { type: object } }
...
security:
  - bearer: []
    session: []
paths:
...
  /api/v1/talk:
    post:
      tags: [talk]
      summary: Say something
      description: >
        Messages sent before any answer are answered together: the reply attaches
        to the newest and the earlier ones carry coveredById. The message is
        stored before reasoning is attempted, so a model outage never costs it.
        Refused for agent keys.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [said]
              properties: { said: { type: string, minLength: 1, maxLength: 8000 } }
      responses:
        '201':
          description: Stored, with a reply when one could be produced
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: '#/components/schemas/Exchange' }
        '403':
          description: An agent key cannot be used for conversation
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    get:
      tags: [talk]
      summary: History
      parameters:
        - { name: day, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: cursor, in: query, schema: { type: string } }
      responses: { '200': { description: List } }
...